moved arcsearch to own endpoint, added renameFolder small bugfix for arc creation (added default value to group id)
